now take username as arg not from config.h

+++ ds.c
@@ -24,6 +24,8 @@
 #include <login_cap.h>
 #include <bsd_auth.h>
 
+#define len(X) (sizeof(X)/sizeof(X[0]))
+
 /* variables */
 static char *progname;
 
@@ -33,29 +35,24 @@ static struct {
 	char *dsp;
 } xserver;
 
-static jmp_buf termdm;
-
 /* configuration */
 #include <config.h>
 
 /* functions */
 static char *cat (char *, char *);
 static void die (const char *, ...);
-static void makedaemon (void);
-static void resetserver (void);
-static void runsession (void);
+static void runsession (struct passwd *);
 static void serverhandler (int);
 static void setsignal (int, void (*) (int));
-static void spawnwm (void);
+static void spawnwm (struct passwd *);
 static void startserver (void);
-static void termhandler (int);
 
 char *
 cat (char *dst, char *src) {
 	char buff[256];
 
-	strncpy(buff,dst,sizeof(buff)/sizeof(buff[0]));
-	strncat(buff,src,sizeof(buff)/sizeof(buff[0]));
+	strncpy(buff, dst, len(buff));
+	strncat(buff, src, len(buff));
 	return strdup(buff);
 }
 
@@ -63,71 +60,30 @@ void
 die (const char *error, ...) {
 	va_list alist;
 
-	va_start(alist,error);
-	vfprintf(stderr,error,alist);
+	va_start(alist, error);
+	vfprintf(stderr, error, alist);
 	va_end(alist);
 
 	exit(EXIT_FAILURE);
 }
 
 void
-makedaemon (void) {
-	pid_t pid;
-	int fd;
-
-	switch ((pid = fork())) {
-		case -1:
-			perror("fork");
-			die("%s: cannot fork to go daemon.\n");
-			break;
-		case 0:
-			setsid();
-			chdir("/");
-			fd = open("/dev/null",O_RDWR);
-			dup2(fd,0);
-			dup2(fd,1);
-			dup2(fd,2);
-			break;
-		default:
-			exit(EXIT_SUCCESS);
-	}
-}
-
-void
-resetserver (void) {
-	unsigned int nwins;
-	Display *dsp;
-	Window *win;
-	Window root;
-
-	if ((dsp = XOpenDisplay(xserver.dsp)) == NULL)
-		die("%s: unable to connect to the x server\n", progname);
-
-	root = RootWindow(dsp, DefaultScreen(dsp));
-	
-	XQueryTree(dsp, root, &root, &root, &win, &nwins);
-	for (; nwins--; win++)
-		XKillClient(dsp, *win);
-
-	XSync(dsp, 0);
-}
-
-void
-runsession (void) {
+runsession (struct passwd *pwd) {
 	pid_t pid;
 
 	switch ((pid = fork())) {
 		case -1:
 			perror("fork");
-			die("%s: cannot fork to run client session.\n",progname);
+			die("%s: cannot fork to run client session.\n", progname);
 			break;
 		case 0:
 			/* drop privileges & spawn wm */
-			spawnwm();
+			printf("dropping privileges and spawning wm\n");
+			spawnwm(pwd);
 			break;
 		default:
 			waitpid(pid,NULL,0);
-			fprintf(stdout,"%s: session finished (pid %d).\n",progname,pid);
+			printf("%s: session finished (pid %d).\n", progname, pid);
 			break;
 	}
 }
@@ -149,26 +105,26 @@ setsignal (int signum, void (*handler) (int)) {
 }
 
 void
-spawnwm (void) {
+spawnwm (struct passwd *pwd) {
 	char *xinit, *cmd[3], *env[7];
-	struct passwd *pwd;
+	int i;
 
-	pwd = getpwnam(user);
-	if (!pwd) {
-		die("Failed to get auth information for user: %s\n", user);
-	}
+	printf("setting user context: ");
 
 	if (setusercontext(NULL, pwd, pwd->pw_uid, LOGIN_SETALL) != 0) {
+		printf("FAIL\n");
 		die("%s: (session process) cannot set user context\n", progname);
+	} else {
+		printf("success\n");
 	}
-
-	chdir(pwd->pw_dir);
-
+	
 	xinit = cat(pwd->pw_dir, "/.xinitrc"); 
 
 	cmd[0] = pwd->pw_shell;
 	cmd[1] = xinit;
 	cmd[2] = NULL;
+
+	printf("xinit command: %s %s\n", cmd[0], cmd[1]);
 
 	env[0] = cat("HOME=", pwd->pw_dir);
 	env[1] = cat("LOGNAME=", pwd->pw_name);
@@ -178,6 +134,14 @@ spawnwm (void) {
 	env[5] = cat("DISPLAY=", xserver.dsp);
 	env[6] = NULL;
 
+	printf("setting env to:\n");
+	for (i = 0; i < len(env) - 1; i++) {
+		printf("\t%s\n", env[i]);
+	}
+	
+	printf("chdir\n");
+	chdir(pwd->pw_dir);
+	printf("execve\n");
 	execve(cmd[0], cmd, env);
 	die("%s: (session process) cannot run session wm.\n",progname);
 }
@@ -187,7 +151,6 @@ startserver (void) {
 	pid_t pid;
 	time_t starttime;
 	char *display_name = ":0";
-	int fd;
 	int i;
 
 	xserver.started = 0;
@@ -200,60 +163,58 @@ startserver (void) {
 		case 0:
 			/* run x server, (TODO: add an auth feature) */
 			setsignal(SIGUSR1,SIG_IGN);
-			fd = open("/dev/null",O_RDWR);
-			dup2(fd,0);
-			dup2(fd,1);
-			dup2(fd,2);
 			execv(xcmd[0],xcmd);
 			die("%s: (server process) server execution failed.\n",progname); /* not seen ... */
 			break;
 		default:
 			xserver.pid = pid;
 			starttime = time(NULL);
+			
 			while (xserver.started == 0) {
 				if (time(NULL) - starttime > xtimeout)
 					die("%s: server startup timed out.\n",progname);
 				usleep(50000);
 			}
-			for (i=0; i<sizeof(xcmd)/sizeof(xcmd[0]); i++)
+
+			for (i=0; i < len(xcmd); i++) {
 				if (xcmd[i] && xcmd[i][0] == ':') {
 					display_name = xcmd[i];
 					break;
 				}
+			}
+
 			xserver.dsp = display_name;
-			fprintf(stdout,"%s: x server started (pid: %d)\n",progname,pid);
+			printf("%s: x server started (dis: %s)"
+			        "(pid: %d)\n", progname, display_name, pid);
 			setsignal(SIGUSR1,SIG_IGN);
 			break;
 	}
 }
 
-void
-termhandler (int sig) {
-	if (sig == SIGTERM)
-		longjmp(termdm,1);
-}
-
 int
 main (int argc, char *argv[]) {
+	struct passwd *pwd;
 	progname = argv[0];
+
+	if (argc != 2) 
+		die("usage: %s username\n", progname);
+	
 	if (geteuid() != 0)
-		die("%s: only root can launch ds.\n",progname);
+		die("%s: only root can launch ds.\n", progname);
 
-	makedaemon();
-	fprintf(stdout,"%s (pid: %d)\n",progname,getpid());
+	if (!(pwd = getpwnam(argv[1]))) 
+		die("Failed to get auth information for user: %s\n", argv[1]);
+
+	printf("%s (pid: %d)\n", progname, getpid());
+
+	printf("start server\n");
 	startserver();
+	printf("runsession\n");
+	runsession(pwd);
 
-	if (!setjmp(termdm)) {
-		setsignal(SIGTERM,termhandler);
-		while (1) {
-			runsession();
-			resetserver();
-			sleep(2);
-		}
-	}
+	printf("killing x server (%d)\n", xserver.pid);
+	kill(xserver.pid, SIGTERM);
 
-	fprintf(stdout,"killing x server (%d)\n",xserver.pid);
-	kill(xserver.pid,SIGTERM);
 	exit(EXIT_SUCCESS);
 	return 0;
 }
